Refactor watson.auth.service out of watson route
