--- api/routes/auth.route/auth.route.js
@@ -1,26 +1,0 @@
-require('dotenv');
-
-const express = require('express');
-const router = express.Router();
-
-router.get('/', (req, res) => {
-  const authorization = new watson.AuthorizationV1({
-    username: process.env.SPEECH_TO_TEXT_USERNAME,
-    password: process.env.SPEECH_TO_TEXT_PASSWORD,
-    url: watson.SpeechToTextV1.URL
-  });
-
-  try {
-    authorization.getToken(function(err, token) {
-      if (!token) {
-        console.log('error:', err);
-      } else {
-        return res.status(200).send(token);
-      }
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-module.exports = router;
